requirements.txt
behave>=1.2.6
behave-html-pretty-formatter>=1.11.1
pylink-square>=1.2.0
prettytable